fix: add float32 rounding barriers to stop arm64 FMA contraction (#59)
The Go compiler contracts float32 a*b+c into a fused multiply-add on
arm64 (FMADDS) but emits separate mul+add on amd64. The reference the
port targets is a -ffp-contract=off FFmpeg build, which never fuses, so
every unguarded float32 a*b+c is both a fidelity mismatch and a
GOARCH-dependent output split.

Add explicit float32() rounding barriers at every contracting float32
site in the encoder: the 3GPP psy model, the LAME FIR, calcBark, the
coder stereo tools, twoloop form factor, the NMR side EMA, and the enc
M/S and IS paths. Verified float32-FMA-clean via
GOARCH=arm64 go build -gcflags=-S ./...

This is a no-op on amd64 (which never contracted) and makes the encoder
produce byte-identical output on arm64 and amd64 given identical input.

+++ internal/coder/twoloop.go
@@ -67,8 +67,8 @@ func FindFormFactor(groupLen, swbSize int, thresh float32, scaled []float32,
 			}
 			variance = fmath.Sqrt32(variance * iswbSizem1)
 
-			e2 = float32(e2 * iswbSize) // rounding barrier: no deferred FMA
-			frm := e / min(e+4*variance, maxval)
+			e2 = float32(e2 * iswbSize)                   // rounding barrier: no deferred FMA
+			frm := e / min(e+float32(4*variance), maxval) // no FMA
 			form += e2 * fmath.Sqrt32(frm) / max(0.5, nzl)
 			weight += e2
 		}
